requirements.txt
python-telegram-bot==20.7
paypalrestsdk==1.13.3
sqlite3
python-dotenv==1.0.0
requests==2.31.0
asyncio
logging